Typo Refactor
- Fix typo for constants package
- Remove unnecessary aliases
