+++ commons-bio/src/main/kotlin/ebi/ac/uk/model/constants/SubmitOperation.kt
@@ -1,4 +1,4 @@
-package ebi.ac.uk.model.constans
+package ebi.ac.uk.model.constants
 
 enum class SubmitOperation {
     CREATE, UPDATE
