+++ commons-bio/src/main/kotlin/ebi/ac/uk/model/extensions/SectionExt.kt
@@ -2,7 +2,7 @@ package ebi.ac.uk.model.extensions
 
 import ebi.ac.uk.model.File
 import ebi.ac.uk.model.Section
-import ebi.ac.uk.model.constans.SectionFields
+import ebi.ac.uk.model.constants.SectionFields
 
 fun Section.allFiles(): List<File> {
     return files.map { it.fold({ file -> listOf(file) }, { table -> table.elements }) }.flatten()
