--- persistence/src/main/kotlin/ac/uk/ebi/biostd/persistence/common/DbAliases.kt
@@ -1,17 +1,0 @@
-package ac.uk.ebi.biostd.persistence.common
-
-import ac.uk.ebi.biostd.persistence.model.Attribute
-import ac.uk.ebi.biostd.persistence.model.AttributeDetail
-import ac.uk.ebi.biostd.persistence.model.File
-import ac.uk.ebi.biostd.persistence.model.Link
-import ac.uk.ebi.biostd.persistence.model.Section
-import ac.uk.ebi.biostd.persistence.model.Submission
-import ac.uk.ebi.biostd.persistence.model.User
-
-typealias SubmissionDb = Submission
-typealias SectionDb = Section
-typealias AttributeDb = Attribute
-typealias AttributeDetailDb = AttributeDetail
-typealias LinkDb = Link
-typealias FileDb = File
-typealias UserDb = User
