+++ persistence/src/main/kotlin/ac/uk/ebi/biostd/persistence/mapping/SubmissionDbMapper.kt
@@ -1,13 +1,14 @@
 package ac.uk.ebi.biostd.persistence.mapping
 
-import ac.uk.ebi.biostd.persistence.common.AttributeDb
-import ac.uk.ebi.biostd.persistence.common.AttributeDetailDb
-import ac.uk.ebi.biostd.persistence.common.FileDb
-import ac.uk.ebi.biostd.persistence.common.LinkDb
+import ac.uk.ebi.biostd.persistence.model.Attribute as AttributeDb
+import ac.uk.ebi.biostd.persistence.model.AttributeDetail as AttributeDetailDb
+import ac.uk.ebi.biostd.persistence.model.File as FileDb
+import ac.uk.ebi.biostd.persistence.model.Link as LinkDb
 import ac.uk.ebi.biostd.persistence.common.NO_TABLE_INDEX
-import ac.uk.ebi.biostd.persistence.common.SectionDb
-import ac.uk.ebi.biostd.persistence.common.SubmissionDb
-import ac.uk.ebi.biostd.persistence.common.UserDb
+import ac.uk.ebi.biostd.persistence.model.Section as SectionDb
+import ac.uk.ebi.biostd.persistence.model.Submission as SubmissionDb
+import ac.uk.ebi.biostd.persistence.model.Tag as TagDb
+import ac.uk.ebi.biostd.persistence.model.User as UserDb
 import ac.uk.ebi.biostd.persistence.model.AccessTag
 import ac.uk.ebi.biostd.persistence.model.Tabular
 import arrow.core.Either
