+++ persistence/src/main/kotlin/ac/uk/ebi/biostd/persistence/repositories/Repositories.kt
@@ -5,6 +5,7 @@ import ac.uk.ebi.biostd.persistence.model.FULL_DATA_GRAPH
 import ac.uk.ebi.biostd.persistence.model.SecurityToken
 import ac.uk.ebi.biostd.persistence.model.Sequence
 import ac.uk.ebi.biostd.persistence.model.Submission
+import ac.uk.ebi.biostd.persistence.model.Tag
 import ac.uk.ebi.biostd.persistence.model.User
 import org.springframework.data.jpa.repository.EntityGraph
 import org.springframework.data.jpa.repository.EntityGraph.EntityGraphType.LOAD
@@ -22,6 +23,10 @@ interface SubmissionDataRepository : JpaRepository<Submission, Long> {
 interface TagsDataRepository : JpaRepository<AccessTag, Long> {
 
     fun findByName(name: String): AccessTag
+}
+
+interface TagsRepository : JpaRepository<Tag, Long> {
+    fun findByName(name: String): Tag
 }
 
 interface SequenceDataRepository : JpaRepository<Sequence, Long> {
