+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/json/serialization/FileJsonSerializer.kt
@@ -7,7 +7,7 @@ import com.fasterxml.jackson.core.JsonGenerator
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.databind.ser.std.StdSerializer
 import ebi.ac.uk.model.File
-import ebi.ac.uk.model.constans.FileFields
+import ebi.ac.uk.model.constants.FileFields
 
 class FileJsonSerializer : StdSerializer<File>(File::class.java) {
