+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/tsv/TsvSerializer.kt
@@ -4,7 +4,7 @@ import ac.uk.ebi.biostd.tsv.deserialization.TsvDeserializer
 import ac.uk.ebi.biostd.tsv.serialization.TsvToStringSerializer
 import ebi.ac.uk.base.splitIgnoringEmpty
 import ebi.ac.uk.model.Submission
-import ebi.ac.uk.model.constans.SUB_SEPARATOR
+import ebi.ac.uk.model.constants.SUB_SEPARATOR
 
 class TsvSerializer(
     private val tsvSerializer: TsvToStringSerializer = TsvToStringSerializer(),
