+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/xml/desirializer/LinkXmlDeserializer.kt
@@ -2,7 +2,7 @@ package ac.uk.ebi.biostd.xml.desirializer
 
 import ac.uk.ebi.biostd.xml.desirializer.common.BaseXmlDeserializer
 import ebi.ac.uk.model.Link
-import ebi.ac.uk.model.constans.LinkFields
+import ebi.ac.uk.model.constants.LinkFields
 import org.w3c.dom.Node
 
 class LinkXmlDeserializer(private val attributeXmlDeserializer: AttributeXmlDeserializer)
