+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/xml/desirializer/SectionXmlDeserializer.kt
@@ -5,9 +5,9 @@ import ebi.ac.uk.model.FilesTable
 import ebi.ac.uk.model.LinksTable
 import ebi.ac.uk.model.Section
 import ebi.ac.uk.model.SectionsTable
-import ebi.ac.uk.model.constans.FileFields
-import ebi.ac.uk.model.constans.LinkFields
-import ebi.ac.uk.model.constans.SectionFields
+import ebi.ac.uk.model.constants.FileFields
+import ebi.ac.uk.model.constants.LinkFields
+import ebi.ac.uk.model.constants.SectionFields
 import org.w3c.dom.Node
 
 class SectionXmlDeserializer(
