+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/xml/serializer/SubmissionSerializer.kt
@@ -8,7 +8,7 @@ import ac.uk.ebi.biostd.xml.common.writeXmlObj
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.dataformat.xml.ser.ToXmlGenerator
 import ebi.ac.uk.model.Submission
-import ebi.ac.uk.model.constans.SubFields
+import ebi.ac.uk.model.constants.SubFields
 
 class SubmissionSerializer : XmlStdSerializer<Submission>(Submission::class.java) {
