+++ serialization/src/main/kotlin/ac/uk/ebi/biostd/xml/serializer/TableSerializer.kt
@@ -5,12 +5,12 @@ import ac.uk.ebi.biostd.xml.common.writeXmlField
 import ac.uk.ebi.biostd.xml.common.writeXmlObj
 import com.fasterxml.jackson.databind.SerializerProvider
 import com.fasterxml.jackson.dataformat.xml.ser.ToXmlGenerator
-import ebi.ac.uk.model.constans.FileFields
+import ebi.ac.uk.model.constants.FileFields
 import ebi.ac.uk.model.FilesTable
-import ebi.ac.uk.model.constans.LinkFields
+import ebi.ac.uk.model.constants.LinkFields
 import ebi.ac.uk.model.LinksTable
-import ebi.ac.uk.model.constans.OtherFields
-import ebi.ac.uk.model.constans.SectionFields
+import ebi.ac.uk.model.constants.OtherFields
+import ebi.ac.uk.model.constants.SectionFields
 import ebi.ac.uk.model.SectionsTable
 import ebi.ac.uk.model.Table
